TyranoRenameProviderに処理移植
