+++ src/server/server.ts
@@ -10,6 +10,7 @@ import {
 } from 'vscode-languageserver/node';
 
 import { TextDocument } from 'vscode-languageserver-textdocument';
+import { TyranoRenameProvider } from "../subscriptions/TyranoRenameProvider";
 
 // Create a connection for the server
 const connection = createConnection(ProposedFeatures.all);
@@ -17,13 +18,16 @@ const connection = createConnection(ProposedFeatures.all);
 // Create a text document manager
 const documents: TextDocuments<TextDocument> = new TextDocuments(TextDocument);
 
+// Create rename provider instance
+const renameProvider = new TyranoRenameProvider();
+
 connection.onInitialize((params: InitializeParams) => {
   const result: InitializeResult = {
     capabilities: {
       textDocumentSync: TextDocumentSyncKind.Incremental,
       renameProvider: true,
       // 他の機能もここで宣言できます
-    }
+    },
   };
   return result;
 });
@@ -35,30 +39,11 @@ connection.onRenameRequest((params: RenameParams): WorkspaceEdit => {
     return { changes: {} };
   }
 
-  // ここでリネーム処理を実装
-  const workspaceEdit: WorkspaceEdit = {
-    changes: {}
-  };
-
-  // 例: 同じドキュメント内の同じ文字列を全て置換
-  const text = document.getText();
-  const pattern = new RegExp(params.position.toString(), 'g');
-  let match;
-
-  while ((match = pattern.exec(text)) !== null) {
-    if (!workspaceEdit.changes![params.textDocument.uri]) {
-      workspaceEdit.changes![params.textDocument.uri] = [];
-    }
-    workspaceEdit.changes![params.textDocument.uri].push({
-      range: {
-        start: document.positionAt(match.index),
-        end: document.positionAt(match.index + match[0].length)
-      },
-      newText: params.newName
-    });
-  }
-
-  return workspaceEdit;
+  return renameProvider.provideRenameEdits(
+    document,
+    params.position,
+    params.newName,
+  );
 });
 
 // Listen on the text document manager and connection
